+++ changelog.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * "Change password" request and validation tests
     * "Delete account" request and validation tests
     * Access to account page browser test
+    * Account page browser tests
     
 ### Changed